==> 1_Introduction.ipynb <==
# Exercise 1. Introduction
## Text, Web and Social Media Analytics Lab

We first import the packages that we will be using. 

from tweepy import OAuthHandler, API, Cursor

We define variables for the keys and tokens needed to use the Twitter API. We then create an authentication handler, where we pass our keys and tokens to identify ourselves. After doing this, we can now pass our credentials to the Twitter API and search for different tweets, which in this case are related to COVID-19. We then print them all.

consumer_key = 'XXXX'
consumer_secret = 'XXXX'
access_key = 'XXXX'
access_secret = 'XXXX'

auth = OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_key, access_secret)
api = API(auth,wait_on_rate_limit=True)

counter = 0
for tweet in Cursor(api.search, q="#covid19", count=5, lang='en', tweet_mode='extended').items():
  print(tweet.full_text)
  counter += 1
  if counter == 5:
    break
  else:
    pass
