==> 2_Preprocessing.ipynb <==
# Exercise 2. Preprocessing
### Text, Web and Social Media Analytics Lab 

In this exercise, we will be using the 20-Newsgroups dataset. This version of the dataset contains about 11k newsgroups posts from 20 different topics. We will do following steps:
- A) Import and examine data
- B) Remove initial text metadata
- C) Remove numbers, punctuation, tabs and convert to lower case with gensim
- D) Remove stop words and short words
- E) Stemming and Lematization

We first import all the necessary libraries.

import pandas as pd
import re
from gensim.parsing.preprocessing import STOPWORDS, strip_tags, strip_numeric, strip_punctuation, strip_multiple_whitespaces, remove_stopwords, strip_short, stem_text
import pickle
import en_core_web_sm
import nltk
nltk.download('stopwords')
from nltk.corpus import stopwords

## Part A: Import and examine data

We read the json file from the URL and we then show the head of the dataframe to understand how our data looks like.

df = pd.read_json('https://raw.githubusercontent.com/selva86/datasets/master/newsgroups.json')
df.head()

We now show more detailed information about our dataset, including the number of entries, as well as the column names, the count of non-nulls and the data type of each column.

df.info()

We now show all the target names available in our dataset as well as the number of records that belong to that target.

df['target_names'].value_counts()

We now print the content of our first record in order to understand how it really looks like and how it is structured.

print(df['content'][0])

After reading the previous text, we print the corresponding target name to corroborate if it makes sense.

print(df['target_names'][0])

## Part B: Remove initial text metadata

In order to remove the text metadata, we first create two lists. The first list contains words that start a specific line of metadata, which we want to remove. The second list contains specific words that do not bring any value to the meaning of the text and that we also want to remove.

starting_words = ['From:', 'Article-I.D.:', 'Organization:', 'Lines:', 'NNTP-Posting-Host:', 'Distribution:', 'Reply-To:', 'X-Newsreader:', 'Expires:', ' -']
single_words = ['Subject:', 'Summary:', 'Keywords:']

We now create a copy of our dataframe and then remove every line that starts with any word from the first list and all the words from the second list that might be found in the content of our dataset. By doing this, we make sure only the text that actually brings value and explains the meaning of the message remains. 

data = df.copy()

for word in starting_words:
  data['content'] = data['content'].apply(lambda content: re.sub(word + '.*\n', '', content, flags=re.IGNORECASE))

for word in single_words:
  data['content'] = data['content'].apply(lambda content: re.sub(word, '', content, flags=re.IGNORECASE))

We print the content of the first record to see how the text looks after removing the metadata.

print(data.loc[0, 'content'])

## Part C: Remove numbers, punctuation, tabs and convert to lower case with gensim

We now perform a few more transformations, like removing all the numeric characters, all the punctuation symbols, all extra whitespaces between words, and we then finally convert the whole text to lowercase. This way only individual words remain, each separated by a single space and in the same format. 

data['content'] = (data['content'].apply(lambda content: strip_numeric(content))
                                  .apply(lambda content: strip_punctuation(content))
                                  .apply(lambda content: strip_multiple_whitespaces(content))
                                  .apply(lambda content: content.lower()))

We now print the content of the first record again to see how the text looks after the previous transformations.

print(data['content'][0])

## Part D: Remove stop words and short words

At the beginning, we imported stopwords from two different libraries. 

Here we show the stopwords from 'gensim', but we first sort these words and then print them to see what they look like. 

print(sorted(STOPWORDS))

Here we print the stopwords from 'nltk', where we can see that it has a lot more words, including stopwords from different languages. 

print(sorted(stopwords.words()))

We perform two more transformations to the content of our dataset. We first remove the stopwords using those from 'gensim' and then we also remove all words shorter than three characters. This removes letters that might have remained from the previous transformations and that do not bring any value.

data['content'] = data['content'].apply(lambda content: remove_stopwords(content)).apply(lambda content: strip_short(content))

We now print the content of the first record as we have done before to see the changes in the text.

print(data['content'][0])

## Part E: Stemming and Lematization

We create a copy of our data, and then apply a stemming transformation to our data, which removes all possible inflections in the words, like prefixes and suffixes, leaving only the stem of each word, as we can later see when we print the content of the first record.

stemmed_data = data.copy()
stemmed_data['content'] = stemmed_data['content'].apply(lambda content: stem_text(content))

print(stemmed_data['content'][0])

We first load Spacy's english model, then create another copy of our data, and then apply lemmatization to the content of our dataset. With lemmatization, we are also removing all inflections from words, like prefixes and suffixes, but leaving the lemma of the word, which compared to the stem from the previous step, a lemma is an actual word. We can see this when we print the content of the first record again.

nlp = en_core_web_sm.load()

lemmatized_data = data.copy()
lemmatized_data['content'] = lemmatized_data['content'].apply(lambda content: ' '.join([token.lemma_ for token in nlp(content)]))

print(lemmatized_data['content'][0])

Here we are just mounting the Google drive so we can access its folders and files.

from google.colab import drive
drive.mount('/content/drive')

We finally save our transformed datasets in Google drive by using 'pickle'. By saving our datasets in a file, we don't have to run all the transformations again every time. We can just load the files back and continue working on them.

pickle.dump(stemmed_data, open('/content/drive/MyDrive/Colab Notebooks/TWSM Analytics Lab/storage/stemmed_data.p', 'wb'))
pickle.dump(lemmatized_data, open('/content/drive/MyDrive/Colab Notebooks/TWSM Analytics Lab/storage/lemmatized_data.p', 'wb'))
