==> 3_TextRepresentation.ipynb <==
# Exercise 3. Text Representation
### Text, Web and Social Media Analytics

In this exercise, we will derive the document representaion of the preprocessed (stemmed) newsgroups dataset. We will use sklearn, as well as gensim to derive the bag of words document representation. We will calculate the following representations for each package: 

- Absolute frequencies
- Relative frequencies
- TF-IDF frequencies
- N-grams

We first import all the libraries we will be using.

import pickle
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.preprocessing import Binarizer
from gensim.corpora import Dictionary
from gensim.models import TfidfModel

We now load the preprocessed dataset from the previous exercise by using the pickle package. We then print the first row to make sure it was loaded correctly. 

stemmed_data = pickle.load(open('/content/drive/MyDrive/Colab Notebooks/TWSM Analytics Lab/storage/stemmed_data.p', 'rb'))

print(stemmed_data.iloc[0])

## Bag-of-Words in Scikit-Learn

Here, we define a CountVectorizer object, which calculates the absolute term frequency; this means that the value of each word is the number of appearances in the document. We also define two parameters, 'max_df' and 'min_df'. These two parameters are set to 0.95 and 0.05 respectively, which specifies to leave out any vocabulary that has a document frequency higher than 95% and lower than 5%.  

After this, we fit and transform our data to generate a matrix representation of our text. We then print the features in the matrix, as well as the shape of the matrix and the first few values to get an idea of how the matrix looks.

count_vectorizer = CountVectorizer(max_df=0.95, min_df=0.05)
count_vectorizer_matrix = count_vectorizer.fit_transform(stemmed_data['content'])

print('Features:\n{}\n'.format(count_vectorizer.get_feature_names()))
print('Matrix Shape:\n{}\n'.format(count_vectorizer_matrix.shape))
print('First Values:\n{}'.format(count_vectorizer_matrix[:5, :5].todense()))

Now, we define a TfidfVectorizer object, which will calculate the relative term frequency of each word; this means the number of times that a word appears in the document is divided by the length of the document. To do this, we define two parameters, 'use_idf' to false, so the inverse-document-frequency reweighting is not calculated, and 'norm' to 'l1', so the value of all the words in the document sums to one. We also set the parameters 'max_df' and 'min_df' like we did last time. 

We then print the feature names, the matrix shape and the first few values to get an idea of how the matrix looks.

tfidf_vectorizer_l1 = TfidfVectorizer(max_df=0.95, min_df=0.05, use_idf=False, norm='l1')
tfidf_vectorizer_l1_matrix = tfidf_vectorizer_l1.fit_transform(stemmed_data['content'])

print('Features:\n{}\n'.format(tfidf_vectorizer_l1.get_feature_names()))
print('Matrix Shape:\n{}\n'.format(tfidf_vectorizer_l1_matrix.shape))
print('First Values:\n{}'.format(tfidf_vectorizer_l1_matrix[:5, :5].todense()))

We now define another TfidVectorizer object, which will calculate the actual TF-IDF frequency; this means that the frequency of each word is multiplied by the logarithm of the total number of documents divided by the total number of documents that contain that word. We set the 'max_df' and 'min_df' parameters like before, and also set 'smooth_idf' to false, so an extra document with all words is not used for value smoothing. 

Then we also print the feature names, the matrix shape and the first few values.

tfidf_vectorizer = TfidfVectorizer(max_df=0.95, min_df=0.05, smooth_idf=False)
tfidf_vectorizer_matrix = tfidf_vectorizer.fit_transform(stemmed_data['content'])

print('Features:\n{}\n'.format(tfidf_vectorizer.get_feature_names()))
print('Matrix Shape:\n{}\n'.format(tfidf_vectorizer_matrix.shape))
print('First Values:\n{}'.format(tfidf_vectorizer_matrix[:5, :5].todense()))

Here, we define a Binarizer object, which performs a one-hot encoding on the matrix that we had from the CountVectorizer. This means that the values for each word represent if the word is present or not in the document by using ones and zeros. 

We then print the shape and first few values to confirm that the encoding is working properly. We can see that now all values different than zero are replaced by ones. 

binarizer = Binarizer()
binarizer_matrix = binarizer.fit_transform(count_vectorizer_matrix)

print('Matrix Shape:\n{}\n'.format(binarizer_matrix.shape))
print('First Values:\n{}'.format(binarizer_matrix[:5, :5].todense()))

We now want to create a quick comparison of the first three methods shown above for the first document. We remove all the columns that have words that do not appear in the document and append all the rows into a single dataframe. We can now see how the values for each word changes from method to method. 

first_doc_cv = pd.DataFrame(count_vectorizer_matrix[0].todense(), columns=count_vectorizer.get_feature_names()).rename(index={0: "Abs. Freq."})
first_doc_tfid_l1 = pd.DataFrame(tfidf_vectorizer_l1_matrix[0].todense(), columns=tfidf_vectorizer_l1.get_feature_names()).rename(index={0: "Rel. Freq."})
first_doc_tfid = pd.DataFrame(tfidf_vectorizer_matrix[0].todense(), columns=tfidf_vectorizer.get_feature_names()).rename(index={0: "TF-IDF"})
first_doc_bin = pd.DataFrame(binarizer_matrix[0].todense(), columns=count_vectorizer.get_feature_names()).rename(index={0: "One-Hot"})
first_doc_df = first_doc_cv.append(first_doc_tfid_l1).append(first_doc_tfid).append(first_doc_bin)
first_doc_df = first_doc_df.loc[:, (first_doc_df != 0).all()]
first_doc_df

## Bag-of-Words using Gensim

First, we tokenize each document and create a list of words for all of the documents, resulting with a list of lists of words.

corpus_gen = [doc.split() for doc in stemmed_data['content']]

print(corpus_gen[0])

Now, we create a Dictionary object where we give it the list of lists that we generated in the previous step. We also define two parameters, 'no_below' and 'no_above', which mean the minimum number of documents and the maximum percentage of documents, that we want a word to be in to be considered in the dictionary. We then print the result to understand how it looks like.

id2word = Dictionary(corpus_gen)
id2word.filter_extremes(no_below=566, no_above=0.95,)

print(id2word)

To understand better how the Dictionary object works, we print some of its attributes. We first print the actual dictionary of key-value pairs, having the word and its id, as well as all the words considered in the dictionary and, finally, the number of documents each word appears in according to its id. 

print(id2word.token2id)
print(id2word.token2id.keys())
print(id2word.dfs)

Here, we create the bag of words were we take into account the absolute frequency of each word for each document, so basically it tells us how many times a certain word appears in a certain document. 

corpus1 = [id2word.doc2bow(doc) for doc in corpus_gen]

print(corpus1[0])

Here, we take the bag of words we created previously and calculate the relative frequency for each word. To do this, we take the absolute frequency from the previous bag of words and divide it into the total number of words in the document. 

corpus2 = [[(token[0], (token[1] / sum(n for _, n in doc))) for token in doc] for doc in corpus1]

print(corpus2[0])

Here, we perform a one-hot encoding, where we write a one, if the word appears in the document, or zero otherwise. 

corpus3 = [[(token[0], 1) for token in doc] for doc in corpus1]

print(corpus3[0])

Now, we actually calculate the TF-IDF, taking into account all the documents where each word appears in. 

tfidf = TfidfModel(dictionary=id2word, normalize=True)
corpus4 = [tfidf[id2word.doc2bow(doc)] for doc in corpus_gen]

print(corpus4[0])

We want to create a dataframe with the previous calculated frequencies to compare it with the frequencies we got from sklearn. In order to do this, we first get the first document and the words that appear in it to use as columns.

features = []
for key1, value1 in id2word.iteritems():
  for key2, value2 in corpus1[0]:
    if key1 == key2:
      features.append(value1)

print(features)

Now we can append all the results for the first document into a single dataframe.

doc1 = pd.DataFrame([[pair[1] for pair in corpus1[0]]], columns=features, index=['Abs. Freq.'])
doc2 = pd.DataFrame([[pair[1] for pair in corpus2[0]]], columns=features, index=['Rel. Freq'])
doc3 = pd.DataFrame([[pair[1] for pair in corpus3[0]]], columns=features, index=['One-Hot'])
doc4 = pd.DataFrame([[pair[1] for pair in corpus4[0]]], columns=features, index=['TF-IDF'])

corpus_df = doc1.append(doc2).append(doc4).append(doc3)
corpus_df

We print the dataframe that we had generated from the sklearn's frequencies for comparison. We can see that all values are the same, except for the TF-IDF frequencies. This is due to a difference in the formula for the IDF used by sklearn and gensim to calculate the weights, which are the following: 

Scikit-Learn: $idf(word) = log(\frac{docs_{total}}{docs_{word}}) + 1$

Gensim: $idf(word) = log_2(\frac{docs_{total}}{docs_{word}})$

first_doc_df

## N-grams

We now create n-grams by using the CountVectorizer class from Scikit-Learn. We only have to pass the parameter 'ngram_range' to define the size of the wanted combinations. In our case, we are using (2, 2), which means we only want bigrams or combinations of only two words. We also define the parameters 'max_df' and 'min_df' like we have done before, but we can see a difference in this case, where only three features (bigrams) appear in more than 5% of the documents. If we remove the parameter 'min_df', we get over 800,000 features. 

n_gram = CountVectorizer(ngram_range=(2,2), max_df=0.95, min_df=0.05)
n_gram_matrix = n_gram.fit_transform(stemmed_data['content'])

print(n_gram.get_feature_names())

We transform the matrix we got from the previous step into a dataframe and print the head. We can see that the weights are the absolute frequency of each pair of words. 

n_gram_df = pd.DataFrame(n_gram_matrix.todense(), columns=n_gram.get_feature_names())
n_gram_df.head()

When we print the maximum for each of the features, we can see that 'article apr' and 'write article' appears a maximum of four times in a single document, while 'edu write' appears two times.

for col in n_gram_df.columns:
  print('{}: {}'.format(col, n_gram_df[col].max()))
