==> 4_TextRepresentation_2.ipynb <==
# Exercise 4. Text Representation (2)
### Text, Web and Social Media Analytics

In this exercise, we will apply the following models to the data from the previous exercise: 

- Word2Vec
- Doc2Vec
- BERT (lemmatized)

At the end of the exercise, we will derive a corpus with each of them which can be used in later tasks such as classification and clustering. 

## Part 0. Preparation

As a first step, we first install the transformers package, which includes state-of-the-art NLP models for TensorFlow and Pytorch.

!pip install transformers

We now import all the necessary packages and libraries that we will be using.

import pickle
import numpy as np
import pandas as pd
from gensim.models import Word2Vec
from scipy.spatial.distance import cosine
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
import tensorflow as tf
import torch
from transformers import BertTokenizer, BertModel
from keras.preprocessing.sequence import pad_sequences
from torch.utils.data import TensorDataset, DataLoader, RandomSampler, SequentialSampler

We load our lemmatized data from the pickle file that we have saved in one of the previous exercises. We also print the first row to see that the data was loaded correctly.

lemmatized_data = pickle.load(open('/content/drive/MyDrive/Colab Notebooks/TWSM Analytics Lab/storage/lemmatized_data.p', 'rb'))

print(lemmatized_data.iloc[0])

After loading the data, we take each document and tokenize it, so we end up with a list of all the words in the document. We print the first record to see how the data looks like.

corpus_gen = lemmatized_data['content'].apply(lambda text: text.split())

print(corpus_gen[0])

## Part 1. Word2Vec

Here we define a new Word2Vec model and train it using the corpus we created in the previous step. We set the 'size' parameter to 100 to define the dimensionality of the word vectors; we also set the 'min_count' parameter, which indicates to ignore all words that appear in less than 566 documents. After that, we save the model.

model = Word2Vec(corpus_gen, size=100, min_count=566)
model.save('word2vec.model')

In order to get a better idea about certain characteristics of the model, we print the features and the number of features.


print('Features:\n{}\n'.format(sorted(model.wv.vocab.keys())))
print('Num. of Features:\n{}'.format(len(model.wv.vocab.keys())))

Here we show the vector representation of the word 'car', which has a dimension of 100 like we defined previously in the 'size' parameter when creating the model.

model.wv['car']

We can also check which words are similar or close in the vector space to a given one, in this case, 'car'.

model.wv.most_similar('car')

We can check the similarity to more than one word, in this case for 'bike' and 'machine', where we want to see the most similar one by defining the 'topn' to one. 

model.wv.most_similar(positive=['bike', 'machine'], topn=1)

After seeing that each word is represented in a vector, we would like to also convert each document to a vector that is the average of all the words. To do this, we go through each word of each document and get its vector representation. After we have the vector for all the words, we average them and get a vector representation for the document. 

embedded_corpus = []

for document in corpus_gen:
  word_embeddings = []
  for word in document:
    try:
      word_embeddings.append(model.wv[word])
    except KeyError:
      continue
  if len(word_embeddings) > 0:
    document_embedding = np.mean(word_embeddings, axis=0)
    embedded_corpus.append(document_embedding)

We then convert this document embedding to a dataframe, where each row is a document and all the columns are an average of the vector values for each word. We then print the head to understand how our dataframe looks like and we also print the shape of the dataframe to understand its structure. 

embeddings_df = pd.DataFrame(embedded_corpus)

print(embeddings_df.head())
print('\nShape: {}'.format(embeddings_df.shape))

We now save our Word2Vec model as a pickle file so we can access it later.

pickle.dump(embeddings_df, open('/content/drive/MyDrive/Colab Notebooks/TWSM Analytics Lab/storage/WordtoVecModel.pkl', 'wb'))

We know that the document embedding that we performed before has the same dimension as the word embedding, since it was made by averaging all the word vectors. This means we can also look for words that have a similar representation as for the document; to do this we take the first document embedding and lookup for similar words by vector. 

first_document = embedded_corpus[0]

model.wv.similar_by_vector(first_document)

We can also look for documents that are similar with each other by calculating the cosine distance. In order to do this, we take one document, in this case, the first one, and calculate the cosine distance with all the other documents, then we get the document who had the least distance. 

min_distance = 1.0
most_similar_index = 0
counter = 0

for document in embedded_corpus:
  distance = cosine(first_document, document)
  if (distance < min_distance) and (distance != 0.0):
    min_distance = distance
    most_similar_index = counter
  counter += 1

most_similar_index, min_distance

However, when we print the document that is supposed to be similar, we see that they are actually quite different. This means that the document representation that we performed by averaging all the word embeddings might not work correctly.

print(lemmatized_data['content'].iloc[0])
print(lemmatized_data['content'].iloc[most_similar_index])

## Part 2. Doc2Vec

We now define a new Doc2Vec model and train it, where we define the 'vector_size' to 100 and the 'min_count' to 566, just like we did with the previous model. However, the different is that instead of giving the model the raw corpus, we have to create tagged documents for it to work properly.

documents = [TaggedDocument(doc, [i]) for i, doc in enumerate(corpus_gen)]
model = Doc2Vec(documents, vector_size=100, min_count=566)

We see that this model creates an embedding for the whole document, instead of for each word in the document. We print here the embedded first document to see how it looks like. 

first_document_embedding = model.docvecs[0]

print(first_document_embedding)

Since we have now a vector representation for each document, we can look for the most similar document by calculating the cosine distance again. By using the same method as in the previous model, we find the document with the least distance to the first document. 

min_distance = 1.0
most_similar_index = 0
counter = 0

for document in model.docvecs.vectors_docs:
  distance = cosine(first_document_embedding, document)
  if (distance < min_distance) and (distance != 0.0):
    min_distance = distance
    most_similar_index = counter
  counter += 1

most_similar_index, min_distance

We then print the first document as well as its closest in the vector space. We can clearly see that both documents are similar, unlike with the previous approach. 

print(lemmatized_data['content'].iloc[0])
print(lemmatized_data['content'].iloc[most_similar_index])

We now convert our embeddings into a dataframe like we did before, where each row is a different document and the columns are the vector representation. 

embeddings_df = pd.DataFrame(model.docvecs.vectors_docs)

print(embeddings_df.head())
print('\nShape: {}'.format(embeddings_df.shape))

We then save our Doc2Vec model to a pickle file which we can use later. 

pickle.dump(embeddings_df, open('/content/drive/MyDrive/Colab Notebooks/TWSM Analytics Lab/storage/DoctoVecModel.pkl', 'wb'))

## Part 3. BERT

We check if our GPU is available and recognized by Tensorflow. 

device_name = tf.test.gpu_device_name()

if device_name != '':
  print('Found GPU at: {}'.format(device_name))
else:
  raise SystemError('GPU device not found')

We then check if our GPU is also available and recognized by PyTorch. 

if torch.cuda.is_available():
  device = torch.device('cuda')
  print('There are %d GPU(s) available.' %torch.cuda.device_count())
  print('We will use the GPU: ', torch.cuda.get_device_name(0))
else:
  print('No GPU available, using the CPU instead.')
  device = torch.device('cpu')

We add the tags '[CLS]' and '[SEP]' at the beginning and at the end of each document respectively, so that the BERT model can identify the task it was pre-trained to do.  

sentences = ['[CLS] ' + query + ' [SEP]' for query in lemmatized_data['content']]

print(sentences[0])

We define a new BertTokenizer object from a pre-trained model and then tokenize each document. We print the first one to see how the BertTokenizer splits the words.

tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')
tokenized_texts = [tokenizer.tokenize(sentence) for sentence in sentences]

print(tokenized_texts[0])

Here we want to see the size of the largest tokenization performed, where we can see that the number of tokens is quite large.

max_length = max([len(document) for document in tokenized_texts])

print(max_length)

We see that the number of tokens generated for each document vary a lot, so here we want to pad each sequence to make them have the same length. Since we have sequences with variable length strings, we set the parameter 'dtype' to 'object', as it is recommended. We then define the parameter 'maxlen' to 512, which means that the maximum length of each sequence will be increased or decreased to this number; we also chose 512, because that is the length of the tensors that the BERT Model is expecting. The 'value' parameter is set to '[PAD]', which means that all the spaces that we are filling will have this string. Finally, both the 'truncating' and 'padding' parameters are set to 'post', which means that the padding and truncating will be performed at the right side of the values. 

sentences_padded = pad_sequences(tokenized_texts, dtype=object, maxlen=512, 
                                 value='[PAD]', truncating='post', padding='post')

We now convert each of the tokens to a numerical id based on the vocabulary of the documents. 

sentences_converted = [tokenizer.convert_tokens_to_ids(token) for token in sentences_padded]

We create mask for each of the sequences we created in the previous step, where we write a one if the id is higher than zero or zero if the id is lower than zero; in other words, the mask consists of ones and zeros depending on the value of the id given. 

masks = []

for seq in sentences_converted:
  seq_mask = [int(i>0) for i in seq]
  masks.append(seq_mask)

We now create two tensors, one from the sentences converted to ids, which we will use as inputs, and another from the masks, which we will also use as masks. 

inputs = torch.LongTensor(sentences_converted)
masks = torch.LongTensor(masks)

We define a new BertModel from its pre-trained setup 'bert-case-uncased'. 

model = BertModel.from_pretrained('bert-base-uncased')

We then send it to the device that we will use to train our model, which in our case, it's the GPU. We can also see the architecture of the neural network that the model uses.

model.to(device)

We now create a TensorDataset using the inputs and masks that we created before. We also define a SequentialSampler, which will take samples from our dataset. Finally, we create a DataLoader with our dataset, sampler, and using a batch size of 16. 

batch_size = 16
prediction_data = TensorDataset(inputs, masks)
prediction_sampler = SequentialSampler(prediction_data)
prediction_dataloader = DataLoader(prediction_data, sampler=prediction_sampler, batch_size=batch_size)

We now loop through the dataloader, which gives batches to the model to be trained with, and we then append the resulting embeddings given by the model to our list of results.

result = []
i = 0

for batch in prediction_dataloader:
    batch = tuple(t.to(device) for t in batch)
    b_input_ids, b_input_mask = batch
    with torch.no_grad():
        outputs = model(b_input_ids)
    embeddings = outputs.pooler_output # CLS embeddings for the batch
    embeddings = embeddings.detach().cpu().numpy()
    result.append(embeddings)
    i = i + 1
print('DONE')

Here we print the length of our result and the shape of each batch to understand how it looks like. We had 11314 documents and we fed the data in batches of 16 documents, which result in 708 total batches to process all the documents. This is also why the first value of the shape of the batch is 16 and the second value is 768, because that's the size of the embedding that the BERT model outputs. 

print('Length of Result: {}'.format(len(result)))
print('Batch Shape: {}'.format(result[0].shape))

We loop through each batch extracting each embedding to then create a dataframe with all of them. We then print the head of the dataframe to understand how it looks. 

final = []

for batch in result:
   for embedding in batch:
      final.append(embedding)

final_df = pd.DataFrame(final)

print(final_df.head())

Finally, we save our results to a pickle file, so we can use this data later on.

pickle.dump(final_df, open('/content/drive/MyDrive/Colab Notebooks/TWSM Analytics Lab/storage/BertModel.pkl', 'wb'))
