==> 5_TextClassification.ipynb <==
# Exercise 5. Text Classification

## Text, Web and Social Media Analytics Lab

In this exercise, we will build multiple classification models for the newsgroup dataset. We will apply the following steps:

- Document representation with TF-IDF, Word2Vec and BERT
- Naive Bayes Classification Model
- Random Forests
- Grid Search
- BERT for Sequence Classification

We first import all the required libraries that we are going to use. 

import pandas as pd
import pickle
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report, accuracy_score
from sklearn.ensemble import RandomForestClassifier

base_directory = '/content/drive/MyDrive/Colab Notebooks/TWSM Analytics Lab/storage/'

## Part A: Document Representation with TF-IDF, Word2Vec and BERT

We load the three datasets that we had previously preprocessed. 

data_stem = pickle.load(open(base_directory + 'stemmed_data.p', 'rb'))
data_w2v = pickle.load(open(base_directory + 'WordtoVecModel.pkl', 'rb'))
data_bert = pickle.load(open(base_directory + 'BertModel.pkl', 'rb'))

We show the head of the stemmed dataset to check its contents.

data_stem.head()

Since the Word2Vector model only includes the document embeddings, we add the 'target' and 'target_names' columns from the stemmed dataset, because we will be needing these two columns for the classification task.

data_w2v = data_w2v.join(data_stem[['target','target_names']], on=data_w2v.index, how='left')

print(data_w2v.head())

The same as we did before, we add the columns 'target' and 'target_names' to the BERT data, since it only included the embeddings as well. 

data_bert = data_bert.join(data_stem[['target', 'target_names']], on=data_bert.index, how='left')

print(data_bert.head())

We decide to use only a specific subset of topics, so we remove all the rows from all the datasets that do not belong to these topics.

topics = ['soc.religion.christian', 'rec.sport.hockey', 'talk.politics.mideast', 'rec.motorcycles']

data_stem = data_stem[data_stem['target_names'].isin(topics)]
data_w2v = data_w2v[data_w2v['target_names'].isin(topics)]
data_bert = data_bert[data_bert['target_names'].isin(topics)]

Since we saw before that the stemmed dataset still included the documents, we decide to calculate the TF-IDF frequency and create a new dataframe with the 'target' and 'target_names' columns additionally. 

tfidf_frequency = TfidfVectorizer(max_df=0.7, min_df=0.1).fit_transform(data_stem['content'])
data_stem2 = pd.DataFrame(tfidf_frequency.toarray()).join(data_stem[['target', 'target_names']], on=data_stem.index, how='left')

print(data_stem2.head())

Before going into training the models, we define a function which splits the data into the independent variables, which are all columns except for the last two, and the target variable, which is the 'target' column. We also define a test size of 20% the size of the whole dataset and a random state so the results can be reproduced in different circumstances.

def text_train(df):
  return train_test_split(df.iloc[:, :-2], df.target, test_size=0.20, random_state=12)

## Part B: Naive Bayes Classification Model

First we split our stemmed data into training and testing set for the independent and target variables. We then define a Naive Bayes model and fit the data to it. We finally predict on our test data. 

docs_train_s, docs_test_s, y_train_s, y_test_s = text_train(data_stem2)

clf = MultinomialNB()
clf.fit(docs_train_s, y_train_s)

y_pred_s = clf.predict(docs_test_s)

We print the score, the accuracy and the classification report, where we actually see that the model did not perform bad at all. According to the f1-score, we can see that the model had some more troubles for class 17, which is 'talk.politics.mideast', but overall it performed quite well.  

print('Training Score: {}'.format(round(clf.score(docs_train_s, y_train_s), 4)))
print('Accuracy: {}'.format(round(accuracy_score(y_pred_s, y_test_s), 4)))
print('Classification Report:\n{}'.format(classification_report(y_pred_s, y_test_s)))

## Part C: Random Forests

We now define a Random Forest classifier and fit the same data we used in the previous model. We then predict on our test data and print the classification report, where we see that this model actually performed even better than the Naive Bayes. There is a slight increase in the average accuracy, but we can see that the f1-scores are all quite high. One thing to notice is that the training score is very high, which means that the model is overfitting to the training data.

clf2 = RandomForestClassifier(random_state=42)
clf2.fit(docs_train_s, y_train_s)

y_pred_s = clf2.predict(docs_test_s)

print('Training Score: {}'.format(round(clf2.score(docs_train_s, y_train_s), 4)))
print(classification_report(y_pred_s, y_test_s))

Here we also define a new Random Forest classifier and train it, but this time with another dataset. So first, we split our Word2Vector data into training and testing set, which we use to train and predict. We then print the classification report, where we can see that the model is performing poorly with an average accuracy of 0.25. We can also see that the training score is also very high, which means that the classifier overfitted and its poor performance on the test set might be happening because of the increased number of features and value ranges of the Word2Vec embeddings.

docs_train_w, docs_test_w, y_train_w, y_test_w = text_train(data_w2v)

clf3 = RandomForestClassifier(random_state=42)
clf3.fit(docs_train_w, y_train_w)

y_pred_w = clf3.predict(docs_test_w)

print('Training Score: {}'.format(round(clf3.score(docs_train_w, y_train_w), 4)))
print(classification_report(y_pred_w, y_test_w))

Like we did before, here we also define a new Random Forst classifier and train it using another dataset. This time we split the BERT dataset into the respective sets, train our model and predict on our test data. We print the classification report, where we see that the model performed better than the previous one, but still worse than the first random forest classifier. The training score is of 100%, which means this model also overfitted; however, the test score is better than the previous one, which might show how powerful the BERT embeddings are. 

docs_train_b, docs_test_b, y_train_b, y_test_b = text_train(data_bert)

clf4 = RandomForestClassifier(random_state=42)
clf4.fit(docs_train_b, y_train_b)

y_pred_b = clf4.predict(docs_test_b)

print('Training Score: {}'.format(round(clf4.score(docs_train_b, y_train_b), 4)))
print(classification_report(y_pred_b, y_test_b))

## Part D: Grid Search

All machine learning models have parameters that can be tuned in order to improve the model's performance. One way to do this is by defining a set of parameters with which several models will be trained with and a comparison of the model's performance is made to select the best-performing parameters.

To do this, we first define a parameter grid with a couple of values for the 'min_samples_leaf' and 'n_estimators' parameters. We then initialize a new Random Forest Classifier, as well as a GridSearchCV object, to which we pass the estimator and the parameter grid in order to train the models.

param_grid = {'min_samples_leaf': [5, 10], 'n_estimators': [3, 5]}

rf = RandomForestClassifier(random_state=42)
grid_search = GridSearchCV(estimator=rf, param_grid=param_grid, cv=10)
grid_search.fit(docs_train_s, y_train_s)

We can check what value combination of the parameters performed the best.

best_params = grid_search.best_params_

print('Best Parameters:\n{}'.format(best_params))

We get the best model/estimator, which is the model that was trained with the previous parameters, and then predict on our test data. We then print the classification report and see that the model performed well, but not as well as previous models. However, in this model we can see that the training and testing score is closer together, which might mean that overfitting is reduced with these parameters.

best_model = grid_search.best_estimator_
y_pred_bm = best_model.predict(docs_test_s)

print('Training Score: {}'.format(round(best_model.score(docs_train_s, y_train_s), 4)))
print(classification_report(y_pred_bm, y_test_s))

Here we print the classification report of the first Random Forest Classifier that we trained for comparison, which performed slightly better than the previous model, but overfitted on the training set.

print('First Random Forest Results')
print('Training Score: {}'.format(round(clf2.score(docs_train_s, y_train_s), 4)))
print(classification_report(y_pred_s, y_test_s))

## Part E: BERT Model

In order to train a BERT model, we have to make sure to have the transformers package, which we install here. 

!pip install transformers

We import some additional libraries that we will be using to train our next model.

import numpy as np
import random
import torch
from torch.utils.data import TensorDataset, DataLoader, SequentialSampler
from torch.nn.functional import softmax
from transformers import BertTokenizer, BertForSequenceClassification, AdamW, BertConfig
from keras.preprocessing.sequence import pad_sequences

We check if we have an available GPU, which we then tell PyTorch to use. 

if torch.cuda.is_available():
    device = torch.device('cuda')
    print('There are %d GPU(s) available.' %torch.cuda.device_count())
    print('We will use the GPU: ', torch.cuda.get_device_name(0))
else:
    print('No GPU available, using the CPU instead.')
    device = torch.device('cpu')

We will first do some preprocessing again, which most was done in the previous exercise, but we need to do some changes to the data for our BERT model. 

We load our lemmatized data from the pickle file that we had saved before. We then filter by the four topics that we had chosen previously and change the target values respectively, so they range from 0 to 3 for a four class classification task. We then print the head of our dataframe to see that the filtering and the changes we done correctly.

lemmatized_data = pickle.load(open(base_directory + 'lemmatized_data.p', 'rb'))

topics = {'soc.religion.christian':0, 'rec.sport.hockey':1, 'talk.politics.mideast':2, 'rec.motorcycles':3}
lemmatized_data = lemmatized_data[lemmatized_data['target_names'].isin(topics.keys())]
lemmatized_data['target'] = lemmatized_data['target_names'].apply(lambda name: topics[name])

lemmatized_data.head()

We add the '[CLS]' and '[SEP]' markers at the beginning and at the end of each sentences respectively, since BERT needs them. We then tokenize each of the sentences and print the first one to check how it looks.

sentences = ['[CLS] ' + query + ' [SEP]' for query in lemmatized_data['content']]

tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')
tokenized_texts = [tokenizer.tokenize(sentence) for sentence in sentences]

print(tokenized_texts[0])

We now perform some padding and truncating to each of our documents. We define the 'maxlen' parameter to 512 since that is the size BERT works with, the padding value as '[PAD]', and define that the padding and truncating should be performed 'post', which means after the actual tokens. We then print the first document after the transformations to see how it looks like, but we just show the first 77 characters, since the rest will only include the '[PAD]' marker.

sentences_padded = pad_sequences(tokenized_texts, dtype=object, maxlen=512, value='[PAD]', truncating='post', padding='post')

print(sentences_padded[0][:77])

Now we convert each of the tokens to its id. We print the first 77 characters of the first document, where we can identify the values for certain markers. 

sentences_converted = [tokenizer.convert_tokens_to_ids(token) for token in sentences_padded]

print(sentences_converted[0][:77])

We also create a mask for each document, where we write a one if the token id is higher than zero, or zero if it's lower. In this case, all '[PAD]' markers will become zeros and every other token will be a one. We do the same print as before to check if this is really happening.

masks = []

for seq in sentences_converted:
  seq_mask = [int(i>0) for i in seq]
  masks.append(seq_mask)
    
print(masks[0][:77])

Now that our data is preprocessed and ready, we must split it into training and testing sets. We split the ids, the masks and the labels the same way we splitted the datasets for the previous models. This way we will be able to compare accurately. We then print the train and test lengths of the data to make sure all the data was split correctly.

inputs_train, inputs_test, labels_train, labels_test = train_test_split(sentences_converted, lemmatized_data['target'], 
                                                                        test_size=0.2, random_state=12)
masks_train, masks_test = train_test_split(masks, test_size=0.2, random_state=12)

print('Train-Test Lengths\nInputs: {} - {}\nMasks: {} - {}\nLabels: {} - {}'.format(len(inputs_train), 
            len(inputs_test), len(masks_train), len(masks_test), len(labels_train), len(labels_test)))

We now convert all of our dataset variables into torch tensors. 

inputs_train = torch.LongTensor(inputs_train)
inputs_test = torch.LongTensor(inputs_test)

labels_train = torch.tensor(labels_train.values)
labels_test = torch.tensor(labels_test.values)

masks_train = torch.LongTensor(masks_train)
masks_test = torch.LongTensor(masks_test)

We define the batch size we want our model to use for the training, as well as create some more variables. We first create a TensorDataset with all of our training set variables, we also create a SequentialSampler with our train data and a DataLoader for our training session.

batch_size = 8

train_data = TensorDataset(inputs_train, masks_train, labels_train)
train_sampler = SequentialSampler(train_data)
train_dataloader = DataLoader(train_data, sampler=train_sampler, batch_size=batch_size)

Before we start the training, we have to change the BERT model configuration, since by default it's configured for a binary classification. In our case, we have four classes, so that's the number of labels we have to define. We then create a new BertForSequenceClassification model with our configurations and send it to our GPU. We also define a new AdamW optimizer for our model.

config = BertConfig.from_pretrained('bert-base-uncased')
config.num_labels = 4

model = BertForSequenceClassification(config)
model.to(device)

optimizer = AdamW(model.parameters(), lr=1e-5)

We first have to set the state of the model to train, before we actually do it. We then decide to train the model for five epochs and then loop through each batch of data from the data loader. Inside each loop, we get the ids, masks and labels that were inside the data loader and send it to our GPU. We then set the gradients to zero and make the forward pass on the model. We get the outputs, from which we can get the loss, then we perform the backward propagation and update the parameters on the optimizer.  

model.train()

for epoch in range(5):

  for batch in train_dataloader:
      
      b_input_ids = batch[0].to(device)
      b_input_masks = batch[1].to(device)
      b_labels = batch[2].to(device)
      
      model.zero_grad()
      
      outputs = model(b_input_ids, attention_mask=b_input_masks, labels=b_labels)
      loss = outputs.loss
      loss.backward()
      optimizer.step()
    
print('Training Done...')

Similarly than with the train data, we also create a TensorDataset with our test dataset variables, as well as a SequentialSampler and a DataLoader. 

test_data = TensorDataset(inputs_test, masks_test, labels_test)
test_sampler = SequentialSampler(test_data)
test_dataloader = DataLoader(test_data, sampler=test_sampler, batch_size=batch_size)

We now change the state of the model to evaluation. We create an empty list for our batch predictions and then tell torch to not update the gradients, since we will be testing and no longer training. For each batch in the test data loader, we get the ids and the masks only. We then pass these to our model for a forward pass and get the outputs. We pass the logits from the outputs to a softmax function to get the probabilities for each class. We then get the class with the higher probability, which will be our predicted class, detach it from the GPU and append it to our batch predictions.

model.eval()

batch_predictions = []

with torch.no_grad():
  
  for batch in test_dataloader:

    b_input_ids = batch[0].to(device)
    b_input_masks = batch[1].to(device) 

    outputs = model(b_input_ids, attention_mask=b_input_masks)

    probs = softmax(outputs.logits, dim=1)
    predicted = torch.max(probs, 1).indices
    predicted_detached = predicted.detach().cpu().numpy()
    batch_predictions.append(predicted_detached)

We then loop through each batch and append each class value to a flat list. We then print the length to make sure everything was executed correctly and indeed it matches with the test size. 

predictions = []

for batch in batch_predictions:
  for value in batch:
    predictions.append(value)

print(len(predictions))

Finally, we now print the classification report, where we can see that the BERT model outperformed any other model that we tried before and by a really surprising difference with an 98% accuracy.

print(classification_report(labels_test.numpy(), np.array(predictions)))
